[ADMINAPI-29] Create and implement endpoints for claim set management (#340)
* Read Claim set by Id implementation

* Add Claim set implementation

* Refactor claim set model. Add required mappers

* Update GetClaimSetNames to return all the claim sets.

* Refactor and move common add or update resources on the claim set to separate class.

* Add resource claim and auth strategy validations to Add claim set

* Implement Edit claim set end point

* Refactor edit claim set

* Add validation on system reserved claim set on edit flow.

* Add delete claim set end point

* More validations on edit and delete claim set operations

* Edit claim set validation on duplicate resources

* Add unit tests for Edit, delete and update claim set commands changes

* Add resource duplication check on Add claim set

* Fix Edit claim set name disassociate vendor applications

* Ported over changes from Saagar's PR for parity

* Update resource claim children description

* Add existing claimset check on GetClaimSetByIdQuery

* Rename methods to be more descriptive.

* Add validation for parent child resource relation

* Update Auth strategy  schema to exclude AuthStrategyId and DisplayName

* Fixed typos. Added ed-fi license to SwaggerExcludeSchemaFilter.cs.

+++ Application/EdFi.Ods.AdminApp.Management/ClaimSetEditor/DeleteClaimSetCommand.cs
@@ -1,10 +1,13 @@
-﻿// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: Apache-2.0
 // Licensed to the Ed-Fi Alliance under one or more agreements.
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
+using System;
 using System.Linq;
+using EdFi.Ods.AdminApp.Management.ErrorHandling;
 using EdFi.Security.DataAccess.Contexts;
+using static EdFi.Ods.AdminApp.Management.ClaimSetEditor.GetClaimSetsByApplicationNameQuery;
 
 namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
 {
@@ -20,6 +23,13 @@ namespace EdFi.Ods.AdminApp.Management.ClaimSetEditor
         public void Execute(IDeleteClaimSetModel claimSet)
         {
             var claimSetToDelete = _context.ClaimSets.Single(x => x.ClaimSetId == claimSet.Id);
+
+            if (DefaultClaimSets.Contains(claimSetToDelete.ClaimSetName) ||
+                        CloudOdsAdminApp.SystemReservedClaimSets.Contains(claimSetToDelete.ClaimSetName))
+            {
+                throw new Exception($"Claim set({claimSetToDelete.ClaimSetName}) is system reserved.Can not be deleted.");
+            }
+
             var resourceClaimsForClaimSetId =
                 _context.ClaimSetResourceClaims.Where(x => x.ClaimSet.ClaimSetId == claimSet.Id).ToList();
             _context.ClaimSetResourceClaims.RemoveRange(resourceClaimsForClaimSetId);
